tests/wpt/web-platform-tests/eventsource/eventsource-url.htm

<!DOCTYPE html>
<html>
  <head>
    <title>EventSource: url</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var url = "resources/message.py",
            source = new EventSource(url)
        assert_equals(source.url.substr(-(url.length)), url)
        source.close()
      })
    </script>
  </body>
</html>
